requirements.txt
# LinkedIn Cat 依赖项
# ===================

# Core dependencies - Selenium 自动化
selenium>=4.0.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
pydash>=7.0.0

# CLI dependencies - 命令行界面
typer[all]>=0.9.0
rich>=13.0.0

# Cache dependencies - 状态管理
diskcache>=5.6.0

# Config dependencies - 配置系统
pydantic>=2.0.0
pyyaml>=6.0

# API dependencies - REST API
requests>=2.28.0
pandas>=2.0.0
